+++ src/hal/motor/motor.h
@@ -3,10 +3,10 @@
 class Motor
 {
 public:
-    Motor(int location_angle, int pin_1, int pin_2, int angle_coef, int angle_speed_coef, int cos_coef, int min_speed, bool debug);
-    void set_velocity(float linear_speed, float angle, float angular_speed);
+    Motor(int location_angle, int pin_1, int pin_2, int angle_coef, int angle_speed_coef, int cos_coef, double min_speed, bool debug);
+    void set_velocity(double linear_speed, double angle, double angular_speed);
     void run();
-    void run(int speed);
+    void run(double speed);
     void stop();
     void debug();
 
@@ -18,7 +18,7 @@ private:
     int _angle_speed_coef;
     int _cos_coef;
     int _rotation_speed; // СКОРОСТЬ ВРАЩЕНИЯ САМОГО МОТОРА
-    int _min_speed;
+    double _min_speed;
 
     bool _debug{};
 };
